+++ src/main/java/com/mishop/main/model/Rol.java
@@ -1,0 +1,6 @@
+package com.mishop.main.model;
+
+public enum Rol {
+    USER,
+    ADMIN
+}
