+++ src/main/resources/application.yaml
@@ -5,7 +5,7 @@ spring:
   datasource:
     url: jdbc:postgresql://localhost:5432/eshop
     username: postgres
-    password: 00159753
+    password: postgres
   jpa:
     hibernate:
       ddl-auto: update
